saving

+++ src/rb/rb_grow_push_back.c
@@ -6,15 +6,15 @@
 /*   By: npineau <npineau@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/10/25 12:02:40 by npineau           #+#    #+#             */
-/*   Updated: 2017/10/25 12:28:12 by npineau          ###   ########.fr       */
+/*   Updated: 2017/11/28 15:36:48 by npineau          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "inc/rb.h"
 
-int	rb_grow_push_back(t_rb *rb, void const *elem)
+size_t	rb_grow_push_back(t_rb *rb, void const *elem)
 {
-	int	check;
+	size_t	check;
 
 	if (rb_full(*rb))
 	{
