+++ src/rb/rb_grow_push_front.c
@@ -6,15 +6,15 @@
 /*   By: npineau <npineau@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/10/25 12:29:14 by npineau           #+#    #+#             */
-/*   Updated: 2017/10/25 12:29:42 by npineau          ###   ########.fr       */
+/*   Updated: 2017/11/28 15:37:00 by npineau          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "inc/rb.h"
 
-int	rb_grow_push_front(t_rb *rb, void const *elem)
+size_t	rb_grow_push_front(t_rb *rb, void const *elem)
 {
-	int	check;
+	size_t	check;
 
 	if (rb_full(*rb))
 	{
